the exist components are responsive now
