+++ frontend/src/components/layout/home/Home.js
@@ -4,17 +4,19 @@ import Cards from "../cards/Cards";
 import TrendingPlaces from "../trendingPlaces/TrendingPlaces";
 import Parks from "../parks/Parks";
 import Hospitals from "../hospitals/Hospitals";
+import Province from "../province/Province";
 import Foods from "../foods/Foods";
 
 const Home = () => {
   return (
     <>
       <Header />
-      <Foods />
+      <Province />
       <Cards />
       <TrendingPlaces />
       <Parks />
       <Hospitals />
+      <Foods />
     </>
   );
 };
